Renamed namespace from AuthServer to IdentityAuthServer. Removed custom db. Removed local mdf reference. Cleaned up folders and renamed. Updated appsettings and launchsettings. Updated nuget packages
