--- AuthServer/Controllers/AuthController.cs
@@ -1,69 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IdentityModel.Tokens.Jwt;
-using System.Linq;
-using System.Security.Claims;
-using System.Text;
-using System.Threading.Tasks;
-using AuthServer.Models;
-using AuthServer.Services;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Configuration;
-using Microsoft.IdentityModel.Tokens;
-
-namespace AuthServer.Controllers
-{
-    [Route("api/[controller]/[action]")]
-    [ApiController]
-    public class AuthController : Controller
-    {
-        private IUserService _userService;
-        private IConfiguration _configuration;
-
-        public AuthController(IUserService userService, IConfiguration configuration)
-        {
-            _userService = userService;
-            _configuration = configuration;
-        }
-
-        [HttpGet]
-        [Authorize]
-        public async Task<ActionResult<IEnumerable<User>>> GetUsers()
-        {
-            return Ok(await _userService.GetUsers());
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> Login([FromBody]Login login)
-        {
-            if (login == null)
-            {
-                return BadRequest("Invalid client request");
-            }
-
-            bool authenticated = await _userService.Authenticate(login);
-
-            if (authenticated)
-            {
-                var secretKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["JWT:Key"]));
-                var signinCredentials = new SigningCredentials(secretKey, SecurityAlgorithms.HmacSha256);
-
-                var tokenOptions = new JwtSecurityToken(
-                    issuer: _configuration["JWT:Issuer"],
-                    audience: _configuration["JWT:Audience"],
-                    claims: new List<Claim>(),
-                    expires: DateTime.Now.AddMinutes(5),
-                    signingCredentials: signinCredentials
-                );
-
-                var tokenString = new JwtSecurityTokenHandler().WriteToken(tokenOptions);
-                return Ok(new { Token = tokenString });
-            }
-            else
-            {
-                return Unauthorized();
-            }
-        }
-    }
-}
