--- AuthServer/Data/dummy.txt
@@ -1,1 +1,0 @@
-﻿// mdf files live in this directory
