--- AuthServer/Models/DataDbContext.cs
@@ -1,18 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace AuthServer.Models
-{
-    public partial class DataDbContext : DbContext
-    {
-        public DataDbContext()
-        {
-        }
-
-        public DataDbContext(DbContextOptions<DataDbContext> options)
-            : base(options)
-        {
-        }
-
-        public virtual DbSet<User> Users { get; set; }
-    }
-}
