--- AuthServer/appsettings.json
@@ -1,19 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "DataConnection": "Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=%CONTENTROOTPATH%\\Data\\data.mdf;Initial Catalog=data;Integrated Security=True;Connect Timeout=30",
-    "IdentityConnection": "Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=%CONTENTROOTPATH%\\Data\\identity.mdf;Initial Catalog=identity;Integrated Security=True;Connect Timeout=30"
-  },
-  "JWT": {
-    "Key": "needalongbase64encodedkey",
-    "Issuer": "http://localhost/",
-    "Audience": "APIClient"
-  }
-}
