+++ IdentityAuthServer.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28721.148
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthServer", "AuthServer\AuthServer.csproj", "{2ABDD6D5-26FE-4532-9B61-1263014800C5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IdentityAuthServer", "IdentityAuthServer\IdentityAuthServer.csproj", "{0B00BE0F-8C2D-432D-86B2-1C5CAE09BC82}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2ABDD6D5-26FE-4532-9B61-1263014800C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2ABDD6D5-26FE-4532-9B61-1263014800C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2ABDD6D5-26FE-4532-9B61-1263014800C5}.Release|Any CPU.ActiveCfg = Debug|Any CPU
-		{2ABDD6D5-26FE-4532-9B61-1263014800C5}.Release|Any CPU.Build.0 = Debug|Any CPU
+		{0B00BE0F-8C2D-432D-86B2-1C5CAE09BC82}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0B00BE0F-8C2D-432D-86B2-1C5CAE09BC82}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0B00BE0F-8C2D-432D-86B2-1C5CAE09BC82}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0B00BE0F-8C2D-432D-86B2-1C5CAE09BC82}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
