+++ IdentityAuthServer/Connected Services/Application Insights/ConnectedService.json
