+++ IdentityAuthServer/Models/IdentityAuthDbContext.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore;
+
+namespace IdentityAuthServer.Models
+{
+    public class IdentityAuthDbContext : IdentityDbContext<AppUser>
+    {
+        public IdentityAuthDbContext(DbContextOptions<IdentityAuthDbContext> options)
+              : base(options)
+        {
+        }
+
+    }
+}
