+++ IdentityAuthServer/Models/UserDetails.cs
@@ -1,14 +1,14 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace AuthServer.Models
+namespace IdentityAuthServer.Models
 {
-    public class Register : Login, IUserDetails
+    public class UserDetails: IUserDetails
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
 
-        [Required]
         [EmailAddress]
         public string Email { get; set; }
+
     }
 }
