+++ IdentityAuthServer/Properties/launchSettings.json
@@ -1,0 +1,11 @@
+﻿{
+  "profiles": {
+    "IdentityAuthServer": {
+      "commandName": "Project",
+      "environmentVariables": {
+        "ASPNETCORE_URLS": "http://localhost:5050;https://localhost:5100",
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      }
+    }
+  }
+}
