+++ IdentityAuthServer/appsettings.json
@@ -1,0 +1,18 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "IdentityConnection": "Data Source=(LocalDB)\\MSSQLLocalDB;Initial Catalog=IdentityAuth;Integrated Security=True;Connect Timeout=30"
+  },
+  "JWT": {
+    "Key": "needalongbase64encodedkey",
+    "Issuer": "IdentityAuthServer",
+    "Audience": "YourClient"
+  }
+}
